+++ api/src/main/scala/io/github/datacatering/datacaterer/api/PlanRun.scala
@@ -167,12 +167,12 @@ trait PlanRun {
    * Create new ICEBERG generation step with only warehouse path and table name. Uses hadoop as the catalog type.
    *
    * @param name      Data source name
-   * @param path      Warehouse path to generated ICEBERG
    * @param tableName Table name for generated ICEBERG
+   * @param path      Warehouse path to generated ICEBERG
    * @return FileBuilder
    */
-  def icebergJava(name: String, path: String, tableName: String): FileBuilder =
-    iceberg(name, path, tableName)
+  def icebergJava(name: String, tableName: String, path: String): FileBuilder =
+    iceberg(name, tableName, path)
 
   /**
    * Create new POSTGRES generation step with connection configuration